Tweak how the sources.list.d file is installed.
Build the new sources.list.d file separately and move it into place
only if it actually changed (or was missing). If we did move a
new file into place, tell the user about it and recommend they
do an apt update. If we didn't change the file, don't bother the
user with a useless message.